# openapi-formatted.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/law-clauses/by-law/{lawDocumentId}": {
            "get": {
                "operationId": "LawClauseController_findByLawDocument",
                "summary": "\u6839\u636e\u6cd5\u89c4ID\u83b7\u53d6\u6761\u6b3e\u5217\u8868",
                "parameters": [
                    {
                        "name": "lawDocumentId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            }
        },
...
        "/api/v1/law-clauses/import": {
            "post": {
                "operationId": "LawClauseController_importExcel",
                "summary": "Excel\u6279\u91cf\u5bfc\u5165",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            }
        },
        "/api/v1/law-clauses/{id}": {
            "get": {
                "operationId": "LawClauseController_findOne",
                "summary": "\u83b7\u53d6\u6cd5\u89c4\u6761\u6b3e\u8be6\u60c5",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            },
            "put": {
                "operationId": "LawClauseController_update",
                "summary": "\u66f4\u65b0\u6cd5\u89c4\u6761\u6b3e",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            },
            "delete": {
                "operationId": "LawClauseController_remove",
                "summary": "\u5220\u9664/\u505c\u7528\u6cd5\u89c4\u6761\u6b3e",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            }
        },
        "/api/v1/law-clauses/batch/enable": {
            "post": {
                "operationId": "LawClauseController_batchEnable",
                "summary": "\u6279\u91cf\u542f\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            }
        },
        "/api/v1/law-clauses/batch/disable": {
            "post": {
                "operationId": "LawClauseController_batchDisable",
                "summary": "\u6279\u91cf\u505c\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            }
        },
        "/api/v1/law-clauses/batch/delete": {
            "post": {
                "operationId": "LawClauseController_batchDelete",
                "summary": "\u6279\u91cf\u5220\u9664",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-clause"
                ]
            }
        },
...
    },
    "info": {
        "title": "\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0 API",
        "description": "\n## \u6982\u8ff0\n\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0\u540e\u7aef\u63a5\u53e3\u6587\u6863 v1.0\n\n## \u8ba4\u8bc1\n\u6240\u6709\u63a5\u53e3\uff08\u9664\u767b\u5f55\u3001\u6ce8\u518c\u3001\u5065\u5eb7\u68c0\u67e5\u5916\uff09\u5747\u9700\u8981 JWT \u8ba4\u8bc1\u3002\n\u8bf7\u5728\u8bf7\u6c42\u5934\u4e2d\u6dfb\u52a0\uff1a`Authorization: Bearer <token>`\n\n## \u54cd\u5e94\u683c\u5f0f\n- \u6210\u529f\uff1a`{ data: T, meta?: { page, pageSize, total, totalPages } }`\n- \u5931\u8d25\uff1a`{ code: string, message: string, traceId: string, details?: any }`\n\n## \u5206\u9875\n- \u9ed8\u8ba4 page=1, pageSize=10\n- \u6700\u5927 pageSize=1000\n    ",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "\u8f93\u5165 JWT Token \u8fdb\u884c\u8ba4\u8bc1"
            }
        },
        "schemas": {
...
            "BatchIdsDto": {
                "type": "object",
                "properties": {
                    "ids": {
                        "description": "ID\u6570\u7ec4",
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    }
                },
                "required": [
                    "ids"
                ]
            },
...
        }
    }
}